Added indexes to collections.
We use `Map` to represent indexes over an arbitrary number of fields.
By default, we index the `primaryKey` which speed up a lot of internals
of mobx-rest since we do a lot of primaryKey lookups.

The added benchmark's numbers, show an staggering improvement:

> using the indexed id x 431,498 ops/sec ±0.98% (81 runs sampled)
> using the unindexed age x 114 ops/sec ±0.83% (72 runs sampled)

(The results may vary in your machine)