Fix optimistic destroys for models in collections
There is a bug when you try to destroy a model that is into a
collection. If you decide to do this optimistically, the request will
not be able to be performed because you can compute the url() because
the models is not present in the collection.

Added a test for this.